+++ docs/changes.rst
@@ -16,11 +16,15 @@ Unreleased.
 
  - Rewrote :attr:`~wand.api.libc` library loader. [:issue:`409`]
  - Added :meth:`Image.brightness_contrast() <wand.image.BaseImage.brightness_contrast>` method.
+ - Added :meth:`Image.mode() <wand.image.BaseImage.mode>` method.
  - Added :meth:`Image.oil_paint() <wand.image.BaseImage.oil_paint>` method.
  - Added :meth:`Image.opaque_paint() <wand.image.BaseImage.opaque_paint>` method.
+ - Added :meth:`Image.polaroid() <wand.image.BaseImage.polaroid>` method.
  - Added :meth:`Image.rotational_blur() <wand.image.BaseImage.rotational_blur>` method.
  - Added :meth:`Image.shear() <wand.image.BaseImage.shear>` method.
  - Added :meth:`Image.sigmoidal_contrast() <wand.image.BaseImage.sigmoidal_contrast>` method.
+ - Added :meth:`Image.stegano() <wand.image.BaseImage.stegano>` method.
+ - Added :meth:`Image.stereogram() <wand.image.Image.stereogram>` class method.
 
 
 .. _changelog-0.5.3: